[Backend](Error Management) Updated error management for prisma and added missing (#133)
* add(readme): updated

* add(associations): updated errors to use Prisma

* del(association test): no longer required test

* add(auth): update error management to match with prisma

* add(events): update and error management with prisma

* fix(event): fixed error code thrown and fixed error message

* add(news): update and add error management with Prisma

* add(roles): updated error management for Prisma

* add(stocks): update and add error management for Prisma

* add(users): updated error management for prisma

* add(association): error check if given president exists when creating association

* fix(code smells)

* fix(review): smells

+++ libs/backend/feature/stock/src/lib/stocks.service.ts
@@ -1,3 +1,4 @@
+import { AssociationRepository, StockLogsRepository, StockRepository } from '@stud-asso/backend/core/repository';
 import {
   CreateStockDto,
   CreateStockLogsDto,
@@ -6,21 +7,30 @@ import {
   StockLogsWithUserDto,
   UpdateStockDto,
 } from '@stud-asso/shared/dtos';
-import { StockLogsRepository, StockRepository } from '@stud-asso/backend/core/repository';
 
 import { Injectable } from '@nestjs/common';
+import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class StocksService {
   constructor(
+    private readonly associationRepository: AssociationRepository,
     private readonly stockRepository: StockRepository,
     private readonly stockLogsRepository: StockLogsRepository
   ) {}
 
   public async create(userId: number, createBaseDto: CreateStockDto): Promise<StockDto> {
-    const createdStock: StockDto = await this.stockRepository.create(createBaseDto);
-    await this.createStocksLogs(createdStock.id, userId, createdStock.count, createdStock.count, 'create');
-    return createdStock;
+    try {
+      const createdStock: StockDto = await this.stockRepository.create(createBaseDto);
+      await this.createStocksLogs(createdStock.id, userId, createdStock.count, createdStock.count, 'create');
+      return createdStock;
+    } catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        if (error.code === 'P2003' && error.meta.field_name === 'association (index)') {
+          throw new Error('Association Not Found');
+        }
+      }
+    }
   }
 
   public async findAll(): Promise<StockDto[]> {
@@ -28,19 +38,35 @@ export class StocksService {
   }
 
   public async findAllAsso(id: number): Promise<StockDto[]> {
-    return this.stockRepository.findAllAsso(id);
+    const asso = await this.associationRepository.findOne(id);
+    if (!asso) {
+      throw new Error('Association Not Found');
+    }
+    return await this.stockRepository.findAllAsso(id);
   }
 
   public async findAllAssoStockLogs(associationId: number): Promise<StockLogsWithUserDto[]> {
-    return this.stockLogsRepository.findAllAssoStockLogs(associationId);
+    const asso = await this.associationRepository.findOne(associationId);
+    if (!asso) {
+      throw new Error('Association Not Found');
+    }
+    return await this.stockLogsRepository.findAllAssoStockLogs(associationId);
   }
 
   public async findSpecificStockLogs(stockId: number): Promise<StockLogsDto[]> {
-    return this.stockLogsRepository.findSpecificStockLogs(stockId);
+    const stock = await this.stockRepository.findOne(stockId);
+    if (!stock) {
+      throw new Error('Stock Not Found');
+    }
+    return await this.stockLogsRepository.findSpecificStockLogs(stockId);
   }
 
   public async findOne(id: number): Promise<StockDto> {
-    return this.stockRepository.findOne(id);
+    const stock = await this.stockRepository.findOne(id);
+    if (!stock) {
+      throw new Error('Stock Not Found');
+    }
+    return stock;
   }
 
   public async update(id: number, userId: number, updateBaseDto: UpdateStockDto): Promise<any> {
@@ -55,6 +81,9 @@ export class StocksService {
 
   public async delete(id: number, userId: number): Promise<any> {
     const stockBeforeDelete = await this.stockRepository.findOne(id);
+    if (!stockBeforeDelete) {
+      throw new Error('Stock Not Found');
+    }
     const deletedStock = await this.stockRepository.delete(id);
     await this.createStocksLogs(id, userId, stockBeforeDelete.count, 0, 'delete');
     return deletedStock;
